- Fixed Data Type Problem - Auto Create Table in RunTime - Remove Json Object Library - Remove Sample Entity Currency

+++ src/main/java/com/sdm/core/database/entity/ILogEntity.java
@@ -10,15 +10,15 @@ package com.sdm.core.database.entity;
  * @author Htoonlin
  */
 public interface ILogEntity {
-    int getVersion();
+    long getVersion();
     
-    void setVersion(int version);
+    void setVersion(long version);
     
-    int getCreatedBy();
+    long getCreatedBy();
 
-    void setCreatedBy(int userId);
+    void setCreatedBy(long userId);
 
-    int getModifiedBy();
+    long getModifiedBy();
 
-    void setModifiedBy(int userId);
+    void setModifiedBy(long userId);
 }
